--- controller/booking/booking.js
@@ -1,6 +1,0 @@
-
-
-const reservationBooking = async (req, res) => {};
-const bookingListBooking = async (req, res) => {};
-
-export { reservationBooking, bookingListBooking };
