+++ controller/booking/bookingListBooking.js
@@ -1,0 +1,3 @@
+const bookingListBooking = async (req, res) => {};
+
+export default bookingListBooking;
