+++ controller/booking/reservationBooking.js
@@ -1,0 +1,3 @@
+const reservationBooking = async (req, res) => {};
+
+export default reservationBooking;
