--- controller/comment/comment.js
@@ -1,6 +1,0 @@
-
-const detailComment = async (req, res) => {};
-// const myPageComment = async (req, res) => {};
-
-
-export { detailComment };
