+++ controller/comment/detailComment.js
@@ -1,0 +1,4 @@
+const detailComment = async (req, res) => {};
+
+
+export default detailComment;
