+++ controller/comment/myPageComment.js
@@ -1,0 +1,3 @@
+const myPageComment = async (req, res) => {};
+
+export default myPageComment;
