+++ controller/room/detailRoom.js
@@ -1,0 +1,4 @@
+const detailRoom = async (req, res) => {};
+
+
+export default detailRoom;
