+++ controller/room/mainRoom.js
@@ -1,0 +1,4 @@
+const mainRoom = async (req, res) => {};
+
+
+export default mainRoom;
