+++ controller/room/reservationRoom.js
@@ -1,0 +1,3 @@
+const reservationRoom = async (req, res) => {};
+
+export default reservationRoom;
