--- controller/room/room.js
@@ -1,8 +1,0 @@
-
-const mainRoom = async (req, res) => {};
-const detailRoom = async (req, res) => {};
-const searchRoom = async (req, res) => {};
-const wishListRoom = async (req, res) => {};
-const reservationRoom = async (req, res) => {};
-
-export { mainRoom, detailRoom, searchRoom, wishListRoom, reservationRoom };
