+++ controller/room/searchRoom.js
@@ -1,0 +1,3 @@
+const searchRoom = async (req, res) => {};
+
+export default searchRoom;
