+++ controller/room/wishListRoom.js
@@ -1,0 +1,4 @@
+const wishListRoom = async (req, res) => {};
+
+
+export default wishListRoom;
