+++ controller/user/myPageUser.js
@@ -1,0 +1,3 @@
+const myPageUser = async (req, res) => {};
+
+export default myPageUser;
