+++ controller/user/signInUser.js
@@ -1,0 +1,3 @@
+const signInUser = async (req, res) => {};
+
+export default signInUser;
