+++ controller/user/signUpUser.js
@@ -1,0 +1,3 @@
+const signUpUser = async (req, res) => {};
+
+export default signUpUser;
