--- controller/user/user.js
@@ -1,6 +1,0 @@
-
-const signInUser = async (req, res) => {};
-const signUpUser = async (req, res) => {};
-const myPageUser = async (req, res) => {};
-
-export { signInUser, signUpUser, myPageUser };
