+++ routes/commentRouter.js
@@ -1,5 +1,7 @@
 import express from "express";
-import { detailComment } from "../controller/comment/comment.js";
+import detailComment from "../controller/comment/detailComment.js";
+import myPageComment from "../controller/comment/myPageComment.js";
+
 
 const commentRouter = express.Router();
